DWF-client/requirements.txt
requests
py-cpuinfo
nvgpu
psutil
hurry.filesize
paramiko

tensorflow==1.15
sklearn
numpy
pandas

gensim